refactor: Plan타입을 ProjectToDoList 타입에서 분리
